# perf(rules): parallelize rule validation