This PR is a direct backport of https://github.com/semgrep/semgrep/commit/f07c4e4

### Test plan

Existing end-to-end tests should validate that user-facing behavior does not change.

The performance improvement was validated by running a semgrep scan on a repository with a single code file using a clone of the `semgrep_rules` repository as the rule source.

After cloning, the `semgrep_rules` repository was modified to remove any YAML files that are not semgrep rules, including `.github`,`stats` and `.pre-commit-config.yaml`.

The results of `hyperfine opengrep scan --config semgrep-rules single/file/path` are:

Without this PR's changes:

```
  Time (mean ± σ):     25.001 s ±  0.897 s    [User: 134.339 s, System: 49.526 s]
  Range (min … max):   24.550 s … 27.525 s    10 runs
```

With this PR's changes:

```
  Time (mean ± σ):     24.894 s ±  0.282 s    [User: 132.791 s, System: 49.099 s]
  Range (min … max):   24.512 s … 25.616 s    10 runs
```

The improvements might be negligible but they could allow to align the file to upstream for easier backport of future fixes or improvements

Signed-off-by: Leyart <leyart@gmail.com>